@@ -12,30 +12,30 @@
     "init-db": "prisma db pull & prisma migrate diff --from-empty --to-schema-datamodel prisma/schema.prisma --script > prisma/migrations/0_init/migration.sql & prisma migrate resolve --applied 0_init & prisma generate"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.x",
-    "@cedric-ruiu/stylelint-config": "^0.x",
-    "@iconify-json/fa6-brands": "^1.1.19",
+    "@antfu/eslint-config": "2.21.1",
+    "@cedric-ruiu/stylelint-config": "0.4.0",
+    "@iconify-json/fa6-brands": "1.1.19",
     "@mermaid-js/mermaid-cli": "^9.x",
-    "@netlify/functions": "^2.7.0",
-    "@nuxt/content": "2.12.1",
-    "@prisma/client": "5.13.0",
-    "@unocss/nuxt": "0.60.4",
-    "@unocss/preset-icons": "0.60.4",
-    "@unocss/preset-typography": "0.60.4",
-    "@unocss/preset-web-fonts": "0.60.4",
-    "eslint": "^9.x",
-    "eslint-config-prettier": "^9.x",
-    "include-media": "^2.x",
-    "nodemailer": "^6.9.14",
-    "nuxt": "3.11.2",
-    "photoswipe": "^5.4.3",
-    "postcss": "^8.x",
-    "postcss-scss": "^4.x",
-    "prettier": "^3.x",
-    "prisma": "5.13.0",
-    "prisma-erd-generator": "^1.x",
-    "sass": "^1.x",
-    "stylelint": "^16.x",
-    "typescript": "^5.x"
+    "@netlify/functions": "2.8.0",
+    "@nuxt/content": "2.13.0",
+    "@prisma/client": "5.16.1",
+    "@unocss/nuxt": "0.61.0",
+    "@unocss/preset-icons": "0.61.0",
+    "@unocss/preset-typography": "0.61.0",
+    "@unocss/preset-web-fonts": "0.61.0",
+    "eslint": "9.5.0",
+    "eslint-config-prettier": "9.1.0",
+    "include-media": "2.0.0",
+    "nodemailer": "6.9.14",
+    "nuxt": "3.12.2",
+    "photoswipe": "5.4.4",
+    "postcss": "8.4.38",
+    "postcss-scss": "4.0.9",
+    "prettier": "3.3.2",
+    "prisma": "5.16.1",
+    "prisma-erd-generator": "1.11.2",
+    "sass": "1.77.6",
+    "stylelint": "16.6.1",
+    "typescript": "5.5.2"
   }
 }
